requirements.txt
click==7.1.2
Flask==1.1.2
Flask-MySQL==1.5.1
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==1.4.6
pandas==1.0.4
PyMySQL==0.9.3
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.17
Werkzeug==1.0.1
XlsxWriter==1.2.9


